Fix the definition of uint32_t
bug=none
test=none
Review URL: http://codereview.chromium.org/7187011

git-svn-id: https://nativeclient-sdk.googlecode.com/svn/trunk/src@915 050acbb0-2703-11df-ab0a-9f3f633ae91d

+++ debugger/base/debug_blob.h
@@ -10,7 +10,7 @@
 // I'm not sure from where to get these definitions,
 // so I defined them here - for now.
 typedef unsigned char uint8_t;
-typedef unsigned long long uint32_t;  // NOLINT
+typedef unsigned int uint32_t;  // NOLINT
 typedef unsigned long long uint64_t;  // NOLINT
 
 /// Class for working with raw binary data.
